+++ src/components/Persona/Persona.tsx
@@ -5,31 +5,9 @@ import {Envelope} from '@gravity-ui/icons';
 import {Icon} from '../Icon';
 import {PersonaWrap} from '../PersonaWrap';
 
-import {getTwoLetters} from './getTwoLetters';
-
-export interface PersonaProps {
-    /** Visible text */
-    text: string;
-    /** Image source */
-    image?: string;
-    /**
-     * Visual appearance (with or without border)
-     * @deprecated Use `hasBorder` prop instead
-     */
-    theme?: 'default' | 'clear';
-    /** Display border */
-    hasBorder?: boolean;
-    /** Avatar appearance */
-    type?: 'person' | 'email' | 'empty';
-    /** Text size */
-    size?: 's' | 'n';
-    /** Handle click on button with cross */
-    onClose?: (text: string) => void;
-    /** Handle click on component itself */
-    onClick?: (text: string) => void;
-    /** Custom CSS class for root element */
-    className?: string;
-}
+import i18n from './i18n';
+import type {PersonaProps} from './types';
+import {extractTextValue, extractTextView, getTwoLetters} from './utils';
 
 export function Persona({
     size = 's',
@@ -41,12 +19,18 @@ export function Persona({
     text,
     image,
     className,
+    style,
 }: PersonaProps) {
+    const textValue = extractTextValue(text);
+    const textView = extractTextView(text);
+    const closeButtonAriaAttributes: React.AriaAttributes = {
+        'aria-label': i18n('label_remove-button', {persona: textValue}),
+    };
     let avatar: React.ReactNode | null;
 
     switch (type) {
         case 'person':
-            avatar = image ? <img alt={''} src={image} /> : <span>{getTwoLetters(text)}</span>;
+            avatar = image ? <img alt={''} src={image} /> : <span>{getTwoLetters(textValue)}</span>;
             break;
         case 'email':
             avatar = <Icon data={Envelope} size={14} />;
@@ -56,17 +40,27 @@ export function Persona({
             break;
     }
 
+    const handleClick = React.useCallback(() => {
+        onClick?.(textValue);
+    }, [textValue, onClick]);
+
+    const handleClose = React.useCallback(() => {
+        onClose?.(textValue);
+    }, [textValue, onClose]);
+
     return (
         <PersonaWrap
             size={size}
             theme={hasBorder ? 'default' : 'clear'}
             isEmpty={type === 'empty'}
-            onClick={onClick && onClick.bind(null, text)}
-            onClose={onClose && onClose.bind(null, text)}
+            onClick={onClick && handleClick}
+            onClose={onClose && handleClose}
             avatar={avatar}
             className={className}
+            style={style}
+            closeButtonAriaAttributes={closeButtonAriaAttributes}
         >
-            {text}
+            {textView}
         </PersonaWrap>
     );
 }