+++ src/components/PersonaWrap/PersonaWrap.tsx
@@ -18,6 +18,8 @@ export interface PersonaWrapProps {
     onClose?: (event: React.MouseEvent) => void;
     onClick?: (event: React.MouseEvent) => void;
     className?: string;
+    style?: React.CSSProperties;
+    closeButtonAriaAttributes?: React.AriaAttributes;
 }
 
 export function PersonaWrap({
@@ -29,11 +31,16 @@ export function PersonaWrap({
     className,
     avatar,
     children,
+    style,
+    closeButtonAriaAttributes,
 }: PersonaWrapProps) {
     const clickable = Boolean(onClick);
     const closeable = Boolean(onClose);
     return (
-        <div className={b({size, theme, clickable, closeable, empty: isEmpty}, className)}>
+        <div
+            className={b({size, theme, clickable, closeable, empty: isEmpty}, className)}
+            style={style}
+        >
             {/* eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions */}
             <div className={b('main')} onClick={onClick}>
                 {avatar && <div className={b('avatar')}>{avatar}</div>}
@@ -41,7 +48,7 @@ export function PersonaWrap({
             </div>
             {onClose && (
                 // eslint-disable-next-line jsx-a11y/click-events-have-key-events, jsx-a11y/no-static-element-interactions
-                <div className={b('close')} onClick={onClose}>
+                <div className={b('close')} onClick={onClose} {...closeButtonAriaAttributes}>
                     <Icon data={Xmark} size={12} />
                 </div>
             )}