+++ src/components/Select/components/SelectList/SelectList.tsx
@@ -3,11 +3,12 @@ import React from 'react';
 import {List} from '../../../List';
 import {SelectQa, selectListBlock} from '../../constants';
 import type {SelectOption, SelectProps} from '../../types';
-import {getOptionsHeight, getPopupItemHeight} from '../../utils';
 import type {FlattenOption} from '../../utils';
+import {GroupTitleItem, getOptionsHeight, getPopupItemHeight} from '../../utils';
 
 import {GroupLabel} from './GroupLabel';
 import {OptionWrap} from './OptionWrap';
+import {SelectLoadingIndicator} from './SelectLoadingIndicator';
 
 import './SelectList.scss';
 
@@ -15,49 +16,83 @@ type SelectListProps = {
     mobile: boolean;
     onOptionClick: (option: FlattenOption) => void;
     renderOption?: SelectProps['renderOption'];
+    renderOptionGroup?: SelectProps['renderOptionGroup'];
     getOptionHeight?: SelectProps['getOptionHeight'];
+    getOptionGroupHeight?: SelectProps['getOptionGroupHeight'];
     size: NonNullable<SelectProps['size']>;
     value: NonNullable<SelectProps['value']>;
     flattenOptions: FlattenOption[];
     multiple?: boolean;
     virtualized?: boolean;
+    loading?: boolean;
 };
+
+const loadingOption = {value: '__SELECT_LIST_ITEM_LOADING__', disabled: true};
 
 export const SelectList = React.forwardRef<List<FlattenOption>, SelectListProps>((props, ref) => {
     const {
         onOptionClick,
         renderOption,
+        renderOptionGroup,
         getOptionHeight,
+        getOptionGroupHeight,
         size,
         flattenOptions,
         value,
         multiple,
         virtualized,
         mobile,
+        loading,
     } = props;
+    const items = React.useMemo(
+        () => (loading ? [...flattenOptions, loadingOption] : flattenOptions),
+        [flattenOptions, loading],
+    );
+
     const optionsHeight = getOptionsHeight({
-        options: flattenOptions,
+        options: items,
         getOptionHeight,
+        getOptionGroupHeight,
         size,
         mobile,
     });
 
     const getItemHeight = React.useCallback(
         (option: FlattenOption, index: number) => {
-            return getPopupItemHeight({getOptionHeight, size, option, index, mobile});
+            return getPopupItemHeight({
+                getOptionHeight,
+                getOptionGroupHeight,
+                size,
+                option,
+                index,
+                mobile,
+            });
         },
-        [getOptionHeight, mobile, size],
+        [getOptionHeight, getOptionGroupHeight, mobile, size],
     );
 
     const renderItem = React.useCallback(
         (option: FlattenOption, _isItemActive: boolean, itemIndex: number) => {
             if ('label' in option) {
-                return <GroupLabel label={option.label} />;
+                const wrappedRenderOptionGroup = renderOptionGroup
+                    ? (optionLocal: GroupTitleItem) => {
+                          return renderOptionGroup(optionLocal, {
+                              itemHeight: getItemHeight(optionLocal, itemIndex),
+                          });
+                      }
+                    : undefined;
+
+                return <GroupLabel option={option} renderOptionGroup={wrappedRenderOptionGroup} />;
+            }
+            if (option.value === loadingOption.value) {
+                return <SelectLoadingIndicator />;
             }
 
             const wrappedRenderOption = renderOption
-                ? (option: SelectOption) => {
-                      return renderOption(option, {itemHeight: getItemHeight(option, itemIndex)});
+                ? (optionLocal: SelectOption) => {
+                      return renderOption(optionLocal, {
+                          itemHeight: getItemHeight(optionLocal, itemIndex),
+                      });
                   }
                 : undefined;
 
@@ -70,7 +105,7 @@ export const SelectList = React.forwardRef<List<FlattenOption>, SelectListProps>
                 />
             );
         },
-        [renderOption, value, multiple, getItemHeight],
+        [renderOption, renderOptionGroup, value, multiple, getItemHeight],
     );
 
     return (
@@ -81,7 +116,7 @@ export const SelectList = React.forwardRef<List<FlattenOption>, SelectListProps>
             itemClassName={selectListBlock('item')}
             itemHeight={getItemHeight}
             itemsHeight={virtualized ? optionsHeight : undefined}
-            items={flattenOptions}
+            items={items}
             filterable={false}
             virtualized={virtualized}
             renderItem={renderItem}