+++ src/components/utils/FocusTrap.tsx
@@ -15,28 +15,25 @@ const focusTrapContext = React.createContext<FocusTrapContext | undefined>(undef
 
 interface FocusTrapProps {
     enabled?: boolean;
+    /** @deprecated Use autoFocus instead */
     disableAutoFocus?: boolean;
+    autoFocus?: boolean;
     children: React.ReactElement;
 }
-export function FocusTrap({children, enabled = true, disableAutoFocus}: FocusTrapProps) {
+export function FocusTrap({
+    children,
+    enabled = true,
+    disableAutoFocus,
+    autoFocus = true,
+}: FocusTrapProps) {
     const nodeRef = React.useRef<HTMLElement | null>(null);
 
-    const setAutoFocusRef = React.useRef(!disableAutoFocus);
+    const setAutoFocusRef = React.useRef(!disableAutoFocus && autoFocus);
     React.useEffect(() => {
-        setAutoFocusRef.current = !disableAutoFocus;
+        setAutoFocusRef.current = !disableAutoFocus && autoFocus;
     });
 
     const focusTrapRef = React.useRef<FocusTrapInstance>();
-    if (!focusTrapRef.current) {
-        focusTrapRef.current = createFocusTrap([], {
-            initialFocus: () => setAutoFocusRef.current && getFocusElement(nodeRef.current),
-            fallbackFocus: () => nodeRef.current!,
-            returnFocusOnDeactivate: false,
-            escapeDeactivates: false,
-            clickOutsideDeactivates: false,
-            allowOutsideClick: true,
-        });
-    }
 
     const containersRef = React.useRef<Record<string, HTMLElement>>({});
     const updateContainerElements = React.useCallback(() => {
@@ -68,8 +65,18 @@ export function FocusTrap({children, enabled = true, disableAutoFocus}: FocusTra
         (node: HTMLElement | null) => {
             if (enabled && node) {
                 nodeRef.current = node;
+                if (!focusTrapRef.current) {
+                    focusTrapRef.current = createFocusTrap([], {
+                        initialFocus: () => setAutoFocusRef.current && getFocusElement(node),
+                        fallbackFocus: () => node,
+                        returnFocusOnDeactivate: false,
+                        escapeDeactivates: false,
+                        clickOutsideDeactivates: false,
+                        allowOutsideClick: true,
+                    });
+                }
                 updateContainerElements();
-                focusTrapRef.current?.activate();
+                focusTrapRef.current.activate();
             } else {
                 focusTrapRef.current?.deactivate();
                 nodeRef.current = null;
@@ -112,11 +119,7 @@ export function useParentFocusTrap() {
     }, [actions, id]);
 }
 
-function getFocusElement(root: HTMLElement | null) {
-    if (root === null) {
-        throw new Error('');
-    }
-
+function getFocusElement(root: HTMLElement) {
     if (
         !(document.activeElement instanceof HTMLElement) ||
         !root.contains(document.activeElement)